+++ src/AceUtilsCli.h
@@ -1,5 +1,5 @@
-#ifndef COMMAND_LINE_INTERFACE_H
-#define COMMAND_LINE_INTERFACE_H
+#ifndef ACE_UTILS_CLI_H
+#define ACE_UTILS_CLI_H
 
 #include "cli/InputLine.h"
 #include "cli/StreamLineReader.h"